+++ quilt/changelog.md
@@ -1,1 +1,2 @@
 - Fixed lock icon resolution
+- Updated quilt to 1.20.4